Flush error logs and process deconfigs when calling tpmMarkFailed
When tpmMarkFailed is passed an error log the error gets committed.
We need to make sure that any callouts associated with that error log
are processed prior to us exiting this function. It is likely that
shortly after returning that the HWAS state of the TPM will be
queried and we want to ensure it is up to date according to any of
the callouts associated with the error log that was committed.

Change-Id: I773b4b58b10b5e0bb6ae4fd016eac73d5e30fcce
RTC: 244857
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111341
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>