Correct record data when ECC data validation fails with correctable error
API updateRecordData updates the record data, with the given data, when and a
correctable error occurs.  This API will also update the original error log,
associated with the ECC data validation fail, based on the success or failure
of the update.  This done to have a consistent handling of this error from
multiple uses.

Also did some miscellaneous cleanup of code - mainly made several traces
consistent with the other traces.

Change-Id: Ic1277d02b68d83467680578fcfdf8b797dfbacff
RTC: 263440
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111156
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>