HBBL TPM: Read/Set TDP Bit During HBBL TPM Ops
TDP (TPM Deconfigure Protect) bit indicates that an issue occurred during
an interaction with TPM. SBE will set that bit if it could not perform
TPM operations. HBBL will read the bit prior to attempting to interact
with the TPM and will set that bit if any of its TPM ops fail. HBBL
also sets a flag in HBBL->HB comm area to indicate that the bit was set
by HBBL.

Change-Id: I27f6f33f6c12ed5767ae01a7ee3701b367febfb5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111507
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>