Fix a bug in the MRW processor for SMPGROUP targets
This commit fixes a bug in the MRW processor where the destination
SMPGROUP target was always set to the same instance as the source,
regardless of the real bus connection (i.e. group A could be connected
to A and B to B, but A couldn't be connected to B, nor B to A).

Change-Id: I5432e0e0e19180f77acaa5518e9730c93f9c3f8e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111570
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>