Enable the VPD ECC logic when a config flag is set
These code changes will compile the VPD ECC algorithm files when the config
flag CONFIG_COMPILE_VPD_ECC_ALGORITHMS is set to 'y'.

Created two files, vpd_ecc_api_no_op.C and vpd_ecc_api_algorithms.C, to
facilitate the conditional compile.  If the config flag is set to 'n', then
the vpd_ecc_api_no_op.C will be compiled which will not run the VPD ECC
algorithms.  If the config flag is set to 'y', then the vpd_ecc_api_algorithms.C
file will be compiled that will call the necessary VPD ECC algorithms.

These two files are tracked by GIT and can be updated as needed and checked
in. The files that contain the VPD ECC algorithm APIs will not be tracked by
GIT and therefore will not be updated if modified. This is accomplished
via a .gitignore file.

A .gitignore file will have a list of the files that have the VPD ECC algorithm
APIs.  This will allow the actual VPD ECC algorithm code to be copied into
the vpd directory but not be checked in nor pushed up to gerrit.

Creating these lines of separation will hopefully catch any user errors and
prevent any unwanted files being pushed to git.

If a programmer wishes to modify the files associated with the VPD ECC
algorithm, there is a commit associated withe those file that will allow the
user to do so.

Change-Id: I41a345c77b9475a37119fe8a5cde3c292718ec77
RTC: 263370
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111577
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>