Switch serial and parallel testcase execution order
The testcases for standalone simics were split up into two buckets because the
tests in the serial can interfere with the parallel tests. However, they
ran first which has downstream negative side effects. Instead of fixing
those issues, just move the serial tests to run after the parallel
tests. After all, the purpose of the two buckets was to not have to
patch up all the potential issues for the tests in the serial bucket.

Change-Id: I4fff4cfabf4eb02956e1d4c3e938ef9f0b21a7b4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111624
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>