Check that eecache entry is valid in eepromPerformOpCache
Prior to this commit we assumed that if the user passed in a
target that was present, that its associated EECACHE entries
would be valid. However in the case of ancillary entries it
is possible that the target was detected as present but we
were unable to read the ancillary data source so the ancillary
entry was marked invalid.

Change-Id: Iccbf30474c57fb0dcabdbbda162bea6b739c6352
CQ: SW516835
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111674
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>