Node Comm V2: Quote Exchange
This commit introduces logic to perform multithreaded quote exchange between
nodes in the system. A node quote is a compilation of secureboot/TPM data
that is used for secureboot verification.

In version 2 of the multinode quote exchange, each node will ask each other
node for a quote and will provide each other node its own quote upon request.
At the end of the exchange, all nodes will have quotes from all other nodes.
All of the quotes are extended into TPM PCR at the end of the exchange.

The exchange between two individual nodes is performed in parallel. A node
with lower ID number will ask for a quote first and then send its quote to
the peer. The peer will perform the sequence in reverse.

Change-Id: I14d5cb878cc06062ab5756e71ba0a467c8e0345d
RTC: 244858
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111772
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>