Secure Boot: Inhibit second verification of payload
On eBMC platforms, Hostboot is verifying the payload twice
(once when processing the MCL version of PHyp, and once
as part of the operation to move the payload to its final
resting position).  This commit introduces an API argument
to verifyAndMovePayload that allows a caller, as a
non-default option, to inhibit the verification.  The eBMC
path then calls the API with the non-standard option to
avoid the 2nd verify.

Change-Id: Id31e5c34ece0c1767eeb932c7e735b8d13676a8f
CQ: SW517832
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111812
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>