DDIMM SPD CRC Checking
Implementing the logic to understand the CRC that
is part of the DDIMM SPD specification.  The logic
can both check the CRC to find corruption, as well
as compute a new CRC.

Also implements a P10 DD1 workaround to repair the
SPD that may have been corrupted using the data
from our EECACHE as the backup copy if it exists.

During presence detection if we detect the eeprom is there
   if the memory type (byte2) is unrecognized
      check CRC
      if crc is bad
         assume spd is corrupted
         force to DDR4
         mark dimm as present to be fixed later
      else use existing path where we treat the dimm as not there
Once we've determined the dimm is present we check SN/PN
 to see if the part was replaced since last boot
   if the memory type (byte2) is unrecognized or we detect a SN/PN mismatch
      check CRC
      if crc is bad
         assume spd is corrupted
         mark dimm as not new so that we don't refresh our cache from hardware
For all non-new dimms present in the system
   Check CRC of SPD hardware
      commit an informational log if a problem is found
   Check CRC of cached data
   If cached data has no errors
      Write previously cached data out to hardware
   Else log visible error and mark the part as missing

Includes an update to the prebuilt eecache to fix
some bad CRC in a portion of the SPD that is modified
by the simics code.

Change-Id: Ief9dd6b9e081aacedd925cdee60023a728d94583
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112136
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>