Add writeLidFileFromOffset PLDM Request API
Previously we had support to read lids from an offset with PLDM
requests but we were missing support to write the lids. This function
is required for supporting PNOR access over PLDM. PNOR sections
will now be represented as lids on the BMC and we need a way to
write things like errorlogs and attribute updates to the BMC's
copy of the lid.

Change-Id: Iac7628113b2c1ca5d8ceb840c0632177e53142a9
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112007
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>