Adapt to split reset controls for clock muxes MUX3/MUX4
From DD1 to DD2 we split the async_reset controls between MUX3
and MUX4 to allow for finer control of those resets. It turns out
that this fine control is not really needed, but we need to make
sure that MUX3 is still controlled along with MUX4.

Just manipulate the new MUX3 control just like we used to manipulate
the combined control for both (which now only controls MUX4) on DD1.
The bit was unused on DD1 so we don't need an EC level check.

Change-Id: Iccf8de32df03e13386c629dfda59e1f114732858
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112186
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131102
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>