Mask cross-node bus FIRs during MPIPL, restore at runtime
In a multi-node MPIPL where the SMP has already been
established, Hostboot instances running within a node
do not have visiblity to other nodes.  If an error occurred
on a cross-node bus, popping a FIR, and PRD tried to isolate, it
would not be able to walk the PEER_TARGET relationships to the
remote end of the bus and call out the bus correctly.  Therefore,
during MPIPL, Hostboot will save the intended FIR masks for cross-
node buses (in IOHS and associated PAUC targets) to
attributes, and mask off specific FIRs bits to prevent PRD from
responding to cross-node bus errors.  When HBRT combines the
nodal targeting instances into a final, system-wide model, it
will then restore the saved off FIR masks and PRD will be able
to handle cross-node bus errors correctly.  During a normal IPL,
the FSP only unmasks the internode bus FIRs after istep 18, but
from then until HBRT starts, PRD is not watching for those
specific internode erros anyway, so nothing needs to be done
in that case.

Change-Id: I4de0f1ae87ee0846277617237ebc47183184e938
RTC: 257492
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112231
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>