Hostboot Standalone: Sign SBE Image
This commit uses the signSbeImage from the ppe repo to sign the
SBE Image that hostboot uses for its standalone simics environment.
Among other things, this adds the .sb_settings section to the SBE Image.
This commit also removes a simics check that was used to avoid issues
with us not having the valid .sb_settings section in the SBE Image.

Change-Id: I07560a67bcd1ff400e652bcac25653fc1aea0112
RTC:208821
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112271
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>