Support New SBE-to-Hostboot-Bootloader Data Structure Version
This commit adds the support to look for a new version of the
SBE-to-HBBL data structure.  If found, the Bootloader code will use the
new fields in this structure. Among these new fields are the minimum
secure version, the measurement seeprom version, and the system's
HW Keys' Hash.

The Bootloader code will still default to zero or imprint/development
values of these fields if the updated SBE-to-Bootloader data structure
is not found.

This commit also traces and logs the measurement seeprom version among
the other security values.

Change-Id: I60228f112c3335e82678db1a54123a8d3e4777c6
RTC:208821
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112302
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>