SMPGROUP support: Deconfigure SMPGROUP targets with NULL PEER_TARGET
Previously, an SMPGROUP (or IOHS) could have an invalid/non-functional
peer target (specified in the PEER_TARGET attribute) for two reasons:
(1) its peer target becomes deconfigured, or (2) the PEER_TARGET
attribute is NULL. Previously, an SMPGROUP/IOHS target itself would
only be deconfigured when its PEER_TARGET is deconfigured, but not
when it simply has a NULL PEER_TARGET.

The SMP hardware procedures rely on the fact that any bus passed to
the HWP that doesn't have a valid functional PEER_TARGET is itself
marked non-functional. This constraint was satisfied in the past
because Hostboot only passed functional IOHS targets with valid
PEER_TARGETs to the HWPs (pbusLinkSvc filtered NULL PEER_TARGETs out
of the HWP's set of IOHSes).

The SMP HWPs take IOHS targets as arguments, even with the
introduction of the new SMPGROUP targets. However, the new SMPGROUPs
allow for the possibility that a functional, valid IOHS will be given
to the HWPs that has two functional SMPGROUP children, one of which
has a NULL PEER_TARGET attribute. pbusLinkSvc can't filter these out
at the IOHS level, and so we violated the HWPs' constraint on bus
targets and their PEER_TARGET attributes.

This commit fixes this situation by deconfiguring all SMPGROUP targets
that have NULL PEER_TARGETs during the host_gard step.

Change-Id: Ibcf43be11efd4cacf2920ff821968757185930ec
RTC: 256842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112357
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>