SMPGROUP support: Fixes for SMPGROUP deconfiguration rule (commit 6/n)
Previously, whenever an SMPGROUP is deconfigured and its parent IOHS
has no more functional SMPGROUP children, the IOHS itself is
deconfigured.

This isn't correct in cases when the IOHS is not in SMP mode. This
commit adds a check so that a deconfigured SMPGROUP will only
deconfigure its parent IOHS if the IOHS is in SMP mode.

Change-Id: I8c8e5958874ced042898322f5dda366a229358a0
RTC: 256842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112598
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>