Push full topology id table down to the SBE
The SBE only has a single chip-scope when booting.  It will set up
a fabric topology id table that works for the local chip during
initial boot, but during a mpipl that setup is inefficient in
some cases.  During mpipl, Hostboot wants to immediately do
cross-chip and cross-node memory accesses, which requires a more
accurate topology table to be setup in the hardware, specifically
within the EX logic.

This change will explicitly push the system-wide table down to
each SBE so that it can be used on any subsequent mpipl reboots
to configure the EX logic.  Note that even though Hostboot is
typically node-scoped, the ATTR_PROC_FABRIC_TOPOLOGY_ID_TABLE
value is set by the HWSV code on the FSP prior to boot so it
will be accurate.

CQ: SW514556
CMVC-Prereq: 1127496
Change-Id: I8c3f605c22215fb98babab042cb4fa10383f4bea
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112837
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>