Update TPM Logs to Reflect SBE and HBBL Extensions
The SBE and HBBL extend certain data to the primary TPM, but do
not log this information into the TPM's log.  This commit updates the
TPM's log with what the SBE and HBBL were supposed to extend before
any other extensions/logs can be made by the hostboot code.

Change-Id: Ic1613d64be4104c7bbb922d46f73f45609c44946
RTC:263474
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/112803
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>