Re-enable Backup TPM
This commit re-enables backup TPM initialization and implements logic
to re-initialize SPI engine and reset TPM locality on MPIPL. This allows
to send commands to the backup TPM that is now initialized in normal
boot process.

The SPI init procedure was changed slightly so that the clock init happens
before the master reset. That's supposed to be a more correct way to do
SPI init.

Change-Id: I1e680736a90b09580cfb6d3c6eddea95ed5553ac
RTC: 269646
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113095
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>