Enable the VPD ECC Algorithms for Denali
This will set the flag COMPILE_VPD_ECC_ALGORITHMS as defined in file
src/usr/vpd/HBconfig to true when compiling for Denali.

Currently the flag COMPILE_VPD_ECC_ALGORITHMS is set to false and will not
compile the VPD ECC algorithm code.  Setting this to true will force the
compilation of these files.

Change-Id: I3bafee650305f6cb1e10015174b802749e4acb91
RTC: 263370
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113046
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>