Ensure all requested pages are removed
If a caller attempts to mm_remove_pages() using an input address
that is not page-aligned, there is a chance that the requested
range will not actually be removed.  The exposure occurs if
the requested size is greater than the distance from the starting
address to the upper page boundary.

For example:  (starting address, size)
 (0x1100,0x800) = OK, 0x1000..0x2000 will be removed
 (0x1C00,0x800) = BAD, 0x2000..0x2400 is ignored

Change-Id: I3f9abb82c10f2435fe75565567c28b03c87d611f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113078
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>