Disable attribute sync for explicit poweroff and dead sbe fails
Normally, as part of any shutdown, Hostboot will sync its
attributes down to the FSP.  This is a problem if the FSP
is unable (or unwilling) to handle these messages.  Since
the boot processor SBE is used to transmit the data, if
that dies we don't want to start a message exchange with
the FSP.  There are also cases where the FSP may choose
to shut us down for some reason.  There is no reason to
waste the time with the sync in the good case there, and
there are also fail cases where the exchange might not
even work.

Change-Id: I07a62e24678c72e843444ad0c7f4f8761bd1bc89
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113417
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>