SMPGROUP support: Only deconfigure peers of SMPGROUP targets (commit 8/n)
Previously, deconfiguring an IOHS target would cause its peer to be
deconfigured. This is wrong because the peer might be a split link and
retain a valid connection via another bus.

This commit changes the code so that only SMPGROUPs deconfigure their
peers when deconfigured. This will indirectly deconfigure an IOHS when
both of its children SMPGROUPs have had their peers deconfigured.

This also requires that we deconfigure SMPGROUPs that have no peer
target, which is implemented in another commit.

Change-Id: Id6a01791a25023ca8e4fc4ca4ca26d5203f1bdfa
RTC: 256842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113422
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>