Add support for DD1 and DD2 P10 Standalone switching for simics
This commit updates the simbuild for hostboot to use the level of simics
that defaults to DD2. However, we still want to run CI jobs for DD1. So,
the env var STANDALONE_TEST_DD1 will allow for that to occur. Developers
will also be able to switch back by exporting that var in their
environments as well.

Change-Id: I01fb38d9f49864e55f546c52258d344d6ccf484b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113506
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>