Handle DIMM replacement better in SPD CRC circumvention
The previous algorithm was assuming that the cache was always
valid, even if the SN/PN matched.  This is a problem if the
DIMMs are replaced since the SPD will get rewritten with the
cached values.

As part of this change, we are removing the support to handle
a CRC error inside the SN/PN area itself.  The new algorithm
is:

if I detect a SN/PN match
- check CRC of HW
- check CRC of CACHE
- if no CRC errors
- - force a refresh from HW (normal behavior)
- else if any CRC errors
- - force a cache refresh from HW
- - this will be a hard failure until the dimm is fixed
      if the HW is bad to begin with

Change-Id: I58a54c8d0313b5282dc080ce2323109c7d99b27f
CQ: SW520492
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113566
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>