Add overload for sendrecv_pldm_request that takes size instead of vector
Currently sendrecv_pldm_request with variable sized requests require
the caller to pass the entire vector to this interface. In reality
only the size of the vector is really needed. This commit adds a
new interface that will allow the caller to just pass a size instead of
a vector.

Change-Id: I7c24a75e731c1a67be352e285e4056eae0fb306c
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113607
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>