Remove BMC semi-permanent HB_VOLATILE GARD_Reconfig handling
The HB_VOLATILE partition is managed by OpenBMC based on IPL type,
i.e. IPL is a fresh IPL or a re-IPL (reboot or reconfig loop).

If OpenBMC fresh IPL the HB_VOLATILE RW and PRSV partitions
will be removed and the RO HB_VOLATILE partition will be used
as a fresh copy (no history will persist in the HB_VOLATILE
partition on a fresh IPL).

If OpenBMC re-IPL (Hostboot Reconfig Loop) then the
HB_VOLATILE partition will NOT be removed which will allow
the PERSIST_MAGIC to be sensed and logic triggered based
on data in the HB_VOLATILE partition.

This trigger for Hostboot to execute clearGardByType for
the GARD_Reconfig records is not necessary any longer.

Hostboot does not need the cleanup anymore (previously
provided by clearGardByType) since OpenBMC clears out the
GARD_Reconfig records based on similar logic of the fresh IPL
versus re-IPL.

Change-Id: I79511387d24a0e7870698c5398aaf9d1fb72b94c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113714
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>