Support mirroring TPM extends to other nodes
Once the cross-node SMP is connected together, any
time Hostboot attempts to extend a measurement into
a primary TPM in the local node, it must attempt to
mirror the measurement to the primary TPM in all
other nodes.  It must never attempt to mirror
poisoning or separator related measurements.

Change-Id: I03a4312a151317bb3a9ced8a9826331b9d554770
RTC: 268650
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113737
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>