Sync EECACHE to alternate PNOR
Since the FSP code has no understanding of the contents of our
EECACHE section in PNOR, they cannot provide any logic to keep
the 2 separate physical PNORs in sync.  The best they can do is
to notice if there is a difference and notify us to refresh our
data.  In order to keep the copies in sync, Hostboot will write
out the contents of the active side into the alternate side at
the end of the IPL.

Change-Id: Icb1cb981c5b14ba32d82b3bad7acbad1055799ec
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114761
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>