Enable testing of alternate PNOR during IPL
The alt-pnor verification was removed during Denali bringup.
It is past time to put it back.  This logic will perform
a basic read operation across the LPC bus to the Serial
Flash Controller logic connected to the alternate boot
processor to ensure that the redundant hardware path to the
PNOR is valid without requiring a full failover boot.

Change-Id: Ic99185cc80361bb9dd7428b913a8c1b67b18805b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114760
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>