Use fapi position attributes for crohostdumpfapiattrs response
When the host recieves a crohostdumpfapiattrs command from cronus,
if hostboot is running we will dump the current fapi attributes
for each target in ordered binary image. For each target entry
we list the target, chip pos, unit pos, and node pos. Previously
we used positioning infomation from the phys path attribute. This
was not matching up with the positioning values that cronus uses.
This change updates the source of these position values to be fapi
attributes that should make more sense to cronus users.

Change-Id: Ib7ef2b4e409bd18badac60145c0295ca1d477d19
CQ: SW519852
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/113914
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>