SMPGROUP support: Fix bug deconfiguring SMPGROUPs with NULL peers (commit 10/n)
This commit fixes a bug in the HWAS logic for deconfiguring SMPGROUPs
with NULL peers.

Previously the code would check whether the PEER_PATH had a path type
of "NA," which is incorrect, because even empty paths have a path type
of "physical."

Now the code will check to see whether the element in the path has a
targeting type of "NA" rather than checking the path type.

Change-Id: I8c58d647560b6e34aa5ff0f6082f2dfd1a6d8b13
RTC: 256842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114055
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>